interview_analyzer.py
import cv2
import numpy as np
import time
from nltk.sentiment.vader import SentimentIntensityAnalyzer
import nltk

# Download VADER lexicon if not already present
try:
    nltk.data.find('sentiment/vader_lexicon.zip')
except LookupError:
    # If the lexicon isn't found, download it.
    # Note: This requires 'nltk' to be installed (which it is via requirements.txt)
    nltk.download('vader_lexicon')

sia = SentimentIntensityAnalyzer()

# --- 1. Facial & Emotion Mock Analysis ---
def mock_facial_analysis(video_input_data) -> dict:
    """
    MOCK function to simulate facial emotion and attention analysis.
    
    In a real application, this would process video frames using a
    pre-trained deep learning model (e.g., using DeepFace or a custom CNN).
    For the project demo, we return realistic placeholder values.
    
    Args:
        video_input_data: Placeholder for a video stream or path.
    """
    # Simulate processing time
    time.sleep(0.5) 
    
    # Simulate a distribution of detected emotions over a short interview period
    emotions = {
        "confidence": round(np.random.uniform(0.65, 0.95), 2),
        "stress": round(np.random.uniform(0.05, 0.30), 2),
        "attention": round(np.random.uniform(0.80, 0.98), 2),
        "dominant_emotion": np.random.choice(["Neutral", "Concentration", "Slight Anxiety"])
    }
    
    # Facial score is based on confidence metric
    return {
        "facial_score": int(emotions["confidence"] * 100),
        "emotions": emotions,
        "feedback": f"Maintain strong eye contact. Dominant state was {emotions['dominant_emotion']}."
    }

# --- 2. Communication and Sentiment Analysis (Voice/Text) ---
def analyze_communication(transcript: str) -> dict:
    """
    Analyzes the transcribed response for sentiment, clarity, and pacing.
    
    In a real system, you would integrate a separate voice tone analysis
    library (like Librosa/DeepSpeech to get pitch and speed).
    """
    if not transcript:
        return {"score": 0, "sentiment": "Neutral", "clarity_feedback": "No response provided."}

    # Analyze sentiment using VADER
    vs = sia.polarity_scores(transcript)
    
    # Calculate a mock communication score based on positive sentiment
    comm_score = int(vs['compound'] * 50 + 50) # Scale compound score (-1 to 1) to (0 to 100)
    
    # Mock analysis for pacing and clarity 
    word_count = len(transcript.split())
    clarity_feedback = "Vocabulary usage was clear and professional."
    if word_count < 10:
        clarity_feedback = "Response was too brief; try to elaborate more on your points."
    elif word_count > 50:
        clarity_feedback = "Ensure brevity and focus on key points. The answer was slightly long-winded."
        
    return {
        "score": min(100, max(0, comm_score)),
        "sentiment": "Positive" if vs['compound'] > 0.1 else ("Negative" if vs['compound'] < -0.1 else "Neutral"),
        "clarity_feedback": clarity_feedback,
        "sentiment_scores": vs
    }

# --- 3. Overall Employability Score Integration ---
def calculate_employability_score(profile_match: float, interview_score: int) -> int:
    """
    Combines profile match (skill gap) and interview performance into a single score.
    """
    # Simple weighted average: 60% Profile Match, 40% Interview Performance
    final_score = (profile_match * 0.6) + (interview_score * 0.4)
    return int(round(final_score, 0))

if __name__ == '__main__':
    # Test Communication Analysis
    test_transcript_positive = "I am highly motivated, confident, and excited to tackle challenging problems with my team. This role perfectly aligns with my career goals."
    test_transcript_negative = "The past project was difficult, and I felt stressed and somewhat uncertain about the final outcome, but I managed."
    
    print("--- Test Communication Analysis (Positive) ---")
    print(analyze_communication(test_transcript_positive))
    
    print("\n--- Test Communication Analysis (Negative) ---")
    print(analyze_communication(test_transcript_negative))
    
    # Test Employability Score
    mock_profile_match = 85.0
    mock_combined_interview_score = 78 
    
    final_score = calculate_employability_score(mock_profile_match, mock_combined_interview_score)
    print(f"\n--- Employability Score ---")
    print(f"Profile Match: {mock_profile_match}%, Interview Performance: {mock_combined_interview_score}")
    print(f"Final Employability Score: {final_score}")